insert bar: refactor the code so chatgpt functionality is external component
